refactor: split compiler into modules and deduplicate helpers
