+++ HISTORY.tsv
@@ -71,3 +71,4 @@ date	commit	files	pass	fail	error	timeout	subtests_pass	subtests_total
 2026-02-19	5d3f57f	1298	190	378	730	0	0	94982	105027
 2026-02-19	983edf5	1298	191	378	729	0	0	94987	105032
 2026-02-19	8529db8	1298	192	379	727	0	0	94991	105044
+2026-02-20	4d58150	1298	194	381	723	0	0	95004	105073
