--- src/compiler.rs
@@ -1,2331 +1,0 @@
-use std::collections::HashMap;
-use std::sync::atomic::{AtomicUsize, Ordering};
-
-static STATE_COUNTER: AtomicUsize = AtomicUsize::new(0);
-
-use crate::ast::{AssignOp, CallArg, Expr, PhaserKind, Stmt, make_anon_sub};
-use crate::opcode::{CompiledCode, CompiledFunction, OpCode};
-use crate::token_kind::TokenKind;
-use crate::value::Value;
-
-pub(crate) struct Compiler {
-    code: CompiledCode,
-    local_map: HashMap<String, u32>,
-    compiled_functions: HashMap<String, CompiledFunction>,
-    current_package: String,
-    tmp_counter: usize,
-}
-
-impl Compiler {
-    pub(crate) fn new() -> Self {
-        Self {
-            code: CompiledCode::new(),
-            local_map: HashMap::new(),
-            compiled_functions: HashMap::new(),
-            current_package: "GLOBAL".to_string(),
-            tmp_counter: 0,
-        }
-    }
-
-    fn alloc_local(&mut self, name: &str) -> u32 {
-        if let Some(&slot) = self.local_map.get(name) {
-            return slot;
-        }
-        let slot = self.code.locals.len() as u32;
-        self.code.locals.push(name.to_string());
-        self.local_map.insert(name.to_string(), slot);
-        slot
-    }
-
-    pub(crate) fn compile(
-        mut self,
-        stmts: &[Stmt],
-    ) -> (CompiledCode, HashMap<String, CompiledFunction>) {
-        // Hoist sub declarations: register all subs before executing the block
-        self.hoist_sub_decls(stmts);
-        for (i, stmt) in stmts.iter().enumerate() {
-            let is_last = i == stmts.len() - 1;
-            if is_last && let Stmt::Expr(expr) = stmt {
-                // Last expression: leave value on stack for EVAL return
-                self.compile_expr(expr);
-                self.code.emit(OpCode::SetTopic);
-                continue;
-            }
-            self.compile_stmt(stmt);
-        }
-        (self.code, self.compiled_functions)
-    }
-
-    fn compile_stmt(&mut self, stmt: &Stmt) {
-        match stmt {
-            Stmt::Expr(expr) => {
-                self.compile_expr(expr);
-                self.code.emit(OpCode::Pop);
-            }
-            Stmt::Block(stmts) => {
-                if Self::has_block_enter_leave_phasers(stmts) {
-                    let idx = self.code.emit(OpCode::BlockScope {
-                        enter_end: 0,
-                        body_end: 0,
-                        end: 0,
-                    });
-                    for s in stmts {
-                        if let Stmt::Phaser {
-                            kind: PhaserKind::Enter,
-                            body,
-                        } = s
-                        {
-                            for inner in body {
-                                self.compile_stmt(inner);
-                            }
-                        }
-                    }
-                    self.code.patch_block_enter_end(idx);
-                    for s in stmts {
-                        match s {
-                            Stmt::Phaser {
-                                kind:
-                                    PhaserKind::Enter
-                                    | PhaserKind::Leave
-                                    | PhaserKind::Keep
-                                    | PhaserKind::Undo,
-                                ..
-                            } => {}
-                            _ => self.compile_stmt(s),
-                        }
-                    }
-                    self.code.patch_block_body_end(idx);
-                    for s in stmts {
-                        if let Stmt::Phaser {
-                            kind: PhaserKind::Leave | PhaserKind::Keep | PhaserKind::Undo,
-                            body,
-                        } = s
-                        {
-                            for inner in body {
-                                self.compile_stmt(inner);
-                            }
-                        }
-                    }
-                    self.code.patch_loop_end(idx);
-                } else if Self::has_let_deep(stmts) {
-                    // Block contains `let` — wrap in LetBlock for save/restore
-                    let idx = self.code.emit(OpCode::LetBlock { body_end: 0 });
-                    for (i, s) in stmts.iter().enumerate() {
-                        let is_last = i == stmts.len() - 1;
-                        if is_last {
-                            if let Stmt::Expr(expr) = s {
-                                self.compile_expr(expr);
-                                self.code.emit(OpCode::SetTopic);
-                            } else {
-                                self.compile_stmt(s);
-                            }
-                        } else {
-                            self.compile_stmt(s);
-                        }
-                    }
-                    self.code.patch_let_block_end(idx);
-                } else {
-                    for s in stmts {
-                        self.compile_stmt(s);
-                    }
-                }
-            }
-            Stmt::Say(exprs) => {
-                for expr in exprs {
-                    self.compile_expr(expr);
-                }
-                self.code.emit(OpCode::Say(exprs.len() as u32));
-            }
-            Stmt::Print(exprs) => {
-                for expr in exprs {
-                    self.compile_expr(expr);
-                }
-                self.code.emit(OpCode::Print(exprs.len() as u32));
-            }
-            Stmt::Note(exprs) => {
-                for expr in exprs {
-                    self.compile_expr(expr);
-                }
-                self.code.emit(OpCode::Note(exprs.len() as u32));
-            }
-            Stmt::VarDecl {
-                name,
-                expr,
-                type_constraint,
-                is_state,
-            } => {
-                self.compile_expr(expr);
-                if let Some(tc) = type_constraint {
-                    let tc_idx = self.code.add_constant(Value::Str(tc.clone()));
-                    self.code.emit(OpCode::TypeCheck(tc_idx));
-                }
-                let slot = self.alloc_local(name);
-                if *is_state {
-                    let key = format!("__state_{}", STATE_COUNTER.fetch_add(1, Ordering::Relaxed));
-                    let key_idx = self.code.add_constant(Value::Str(key.clone()));
-                    self.code.state_locals.push((slot as usize, key.clone()));
-                    self.code.emit(OpCode::StateVarInit(slot, key_idx));
-                } else {
-                    self.code.emit(OpCode::SetLocal(slot));
-                }
-            }
-            Stmt::Assign {
-                name,
-                expr,
-                op: AssignOp::Assign | AssignOp::Bind,
-            } if name != "*PID" => {
-                self.compile_expr(expr);
-                if let Some(&slot) = self.local_map.get(name.as_str()) {
-                    self.code.emit(OpCode::SetLocal(slot));
-                } else {
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    self.code.emit(OpCode::SetGlobal(name_idx));
-                }
-            }
-            Stmt::If {
-                cond,
-                then_branch,
-                else_branch,
-            } => {
-                self.compile_expr(cond);
-                let jump_else = self.code.emit(OpCode::JumpIfFalse(0));
-                for s in then_branch {
-                    self.compile_stmt(s);
-                }
-                if else_branch.is_empty() {
-                    self.code.patch_jump(jump_else);
-                } else {
-                    let jump_end = self.code.emit(OpCode::Jump(0));
-                    self.code.patch_jump(jump_else);
-                    for s in else_branch {
-                        self.compile_stmt(s);
-                    }
-                    self.code.patch_jump(jump_end);
-                }
-            }
-            Stmt::While { cond, body, label } => {
-                let (pre_stmts, loop_body, post_stmts) = self.expand_loop_phasers(body);
-                for s in &pre_stmts {
-                    self.compile_stmt(s);
-                }
-                let loop_idx = self.code.emit(OpCode::WhileLoop {
-                    cond_end: 0,
-                    body_end: 0,
-                    label: label.clone(),
-                });
-                self.compile_expr(cond);
-                self.code.patch_while_cond_end(loop_idx);
-                for s in &loop_body {
-                    self.compile_stmt(s);
-                }
-                self.code.patch_loop_end(loop_idx);
-                for s in &post_stmts {
-                    self.compile_stmt(s);
-                }
-            }
-            Stmt::For {
-                iterable,
-                param,
-                params,
-                body,
-                label,
-            } => {
-                let (pre_stmts, mut loop_body, post_stmts) = self.expand_loop_phasers(body);
-                for s in &pre_stmts {
-                    self.compile_stmt(s);
-                }
-                // When there's a single named param (-> $k), store its name as a constant
-                // so the VM can bind $k directly without overriding $_
-                let param_idx = param
-                    .as_ref()
-                    .map(|p| self.code.add_constant(Value::Str(p.clone())));
-                let mut bind_stmts = Vec::new();
-                // When there's a single named param, the VM handles binding directly
-                // Only add $_ -> $param assignment for the legacy path (no param_idx)
-                if param.is_some() && param_idx.is_none() {
-                    bind_stmts.push(Stmt::Assign {
-                        name: param.as_ref().unwrap().clone(),
-                        expr: Expr::Var("_".to_string()),
-                        op: AssignOp::Assign,
-                    });
-                }
-                for (i, p) in params.iter().enumerate() {
-                    bind_stmts.push(Stmt::Assign {
-                        name: p.clone(),
-                        expr: Expr::Index {
-                            target: Box::new(Expr::Var("_".to_string())),
-                            index: Box::new(Expr::Literal(Value::Int(i as i64))),
-                        },
-                        op: AssignOp::Assign,
-                    });
-                }
-                if !bind_stmts.is_empty() {
-                    let mut merged = bind_stmts;
-                    merged.extend(loop_body);
-                    loop_body = merged;
-                }
-                let arity = if !params.is_empty() {
-                    params.len() as u32
-                } else {
-                    1
-                };
-                self.compile_expr(iterable);
-                let loop_idx = self.code.emit(OpCode::ForLoop {
-                    param_idx,
-                    param_local: None,
-                    body_end: 0,
-                    label: label.clone(),
-                    arity,
-                    collect: false,
-                });
-                for s in &loop_body {
-                    self.compile_stmt(s);
-                }
-                self.code.patch_loop_end(loop_idx);
-                for s in &post_stmts {
-                    self.compile_stmt(s);
-                }
-            }
-            // C-style loop (non-repeat, no phasers)
-            Stmt::Loop {
-                init,
-                cond,
-                step,
-                body,
-                repeat,
-                label,
-            } if !*repeat => {
-                let (pre_stmts, loop_body, post_stmts) = self.expand_loop_phasers(body);
-                // Compile init statement (if any) before the loop opcode
-                if let Some(init_stmt) = init {
-                    self.compile_stmt(init_stmt);
-                }
-                for s in &pre_stmts {
-                    self.compile_stmt(s);
-                }
-                // Layout: [CStyleLoop] [cond..] [body..] [step..]
-                let loop_idx = self.code.emit(OpCode::CStyleLoop {
-                    cond_end: 0,
-                    step_start: 0,
-                    body_end: 0,
-                    label: label.clone(),
-                });
-                // Compile condition (or push True if none)
-                if let Some(cond_expr) = cond {
-                    self.compile_expr(cond_expr);
-                } else {
-                    self.code.emit(OpCode::LoadTrue);
-                }
-                self.code.patch_cstyle_cond_end(loop_idx);
-                // Compile body
-                for s in &loop_body {
-                    self.compile_stmt(s);
-                }
-                self.code.patch_cstyle_step_start(loop_idx);
-                // Compile step (if any)
-                if let Some(step_expr) = step {
-                    self.compile_expr(step_expr);
-                    self.code.emit(OpCode::Pop);
-                }
-                self.code.patch_loop_end(loop_idx);
-                for s in &post_stmts {
-                    self.compile_stmt(s);
-                }
-            }
-            Stmt::Call { name, args } => {
-                let rewritten_args = Self::rewrite_stmt_call_args(name, args);
-                let positional_only = rewritten_args
-                    .iter()
-                    .all(|arg| matches!(arg, CallArg::Positional(_)));
-
-                // Normalize mutating/structural call statements through Expr::Call
-                // so they reuse call rewrites and method-based mutation paths.
-                if positional_only && Self::is_normalized_stmt_call_name(name) {
-                    let expr_args: Vec<Expr> = rewritten_args
-                        .iter()
-                        .filter_map(|arg| match arg {
-                            CallArg::Positional(expr) => Some(expr.clone()),
-                            _ => None,
-                        })
-                        .collect();
-                    let call_expr = Expr::Call {
-                        name: name.clone(),
-                        args: expr_args,
-                    };
-                    self.compile_expr(&call_expr);
-                    self.code.emit(OpCode::Pop);
-                    return;
-                }
-
-                // Statement-level call: compile positional args only.
-                // Fall back if named args or raw-expression args remain.
-                if positional_only
-                    && rewritten_args
-                        .iter()
-                        .all(|arg| matches!(arg, CallArg::Positional(_)))
-                {
-                    let arity = rewritten_args.len() as u32;
-                    for arg in &rewritten_args {
-                        if let CallArg::Positional(expr) = arg {
-                            self.compile_expr(expr);
-                        }
-                    }
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    self.code.emit(OpCode::ExecCall { name_idx, arity });
-                    return;
-                }
-
-                // Statement-level call with named args: compile values and encode
-                // named args as Pair(name => value), then dispatch without stmt_pool.
-                for arg in &rewritten_args {
-                    match arg {
-                        CallArg::Positional(expr) => self.compile_expr(expr),
-                        CallArg::Named {
-                            name,
-                            value: Some(expr),
-                        } => {
-                            self.compile_expr(&Expr::Literal(Value::Str(name.clone())));
-                            self.compile_expr(expr);
-                            self.code.emit(OpCode::MakePair);
-                        }
-                        CallArg::Named { name, value: None } => {
-                            self.compile_expr(&Expr::Literal(Value::Str(name.clone())));
-                            self.compile_expr(&Expr::Literal(Value::Bool(true)));
-                            self.code.emit(OpCode::MakePair);
-                        }
-                    }
-                }
-                let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                self.code.emit(OpCode::ExecCallPairs {
-                    name_idx,
-                    arity: rewritten_args.len() as u32,
-                });
-            }
-            // Loop control
-            Stmt::Last(label) => {
-                self.code.emit(OpCode::Last(label.clone()));
-            }
-            Stmt::Next(label) => {
-                self.code.emit(OpCode::Next(label.clone()));
-            }
-            Stmt::Redo(label) => {
-                self.code.emit(OpCode::Redo(label.clone()));
-            }
-            Stmt::Return(expr) => {
-                self.compile_expr(expr);
-                self.code.emit(OpCode::Return);
-            }
-            Stmt::Die(expr) => {
-                self.compile_expr(expr);
-                self.code.emit(OpCode::Die);
-            }
-            Stmt::Fail(expr) => {
-                self.compile_expr(expr);
-                self.code.emit(OpCode::Fail);
-            }
-            Stmt::Proceed => {
-                self.code.emit(OpCode::Proceed);
-            }
-            Stmt::Succeed => {
-                self.code.emit(OpCode::Succeed);
-            }
-            // MatchAssign (~~=): coerce value to string
-            Stmt::Assign {
-                name,
-                expr,
-                op: AssignOp::MatchAssign,
-            } if name != "*PID" => {
-                self.compile_expr(expr);
-                self.code.emit(OpCode::StrCoerce);
-                if let Some(&slot) = self.local_map.get(name.as_str()) {
-                    self.code.emit(OpCode::SetLocal(slot));
-                } else {
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    self.code.emit(OpCode::SetGlobal(name_idx));
-                }
-            }
-            Stmt::Assign { .. } => {
-                self.code.emit(OpCode::AssignReadOnly);
-            }
-            // Given/When/Default
-            Stmt::Given { topic, body } => {
-                self.compile_expr(topic);
-                let given_idx = self.code.emit(OpCode::Given { body_end: 0 });
-                for s in body {
-                    self.compile_stmt(s);
-                }
-                self.code.patch_body_end(given_idx);
-            }
-            Stmt::When { cond, body } => {
-                self.compile_expr(cond);
-                let when_idx = self.code.emit(OpCode::When { body_end: 0 });
-                for (i, s) in body.iter().enumerate() {
-                    let is_last = i == body.len() - 1;
-                    if is_last {
-                        if let Stmt::Expr(expr) = s {
-                            self.compile_expr(expr);
-                        } else {
-                            self.compile_stmt(s);
-                        }
-                    } else {
-                        self.compile_stmt(s);
-                    }
-                }
-                self.code.patch_body_end(when_idx);
-            }
-            Stmt::Default(body) => {
-                let default_idx = self.code.emit(OpCode::Default { body_end: 0 });
-                for (i, s) in body.iter().enumerate() {
-                    let is_last = i == body.len() - 1;
-                    if is_last {
-                        if let Stmt::Expr(expr) = s {
-                            self.compile_expr(expr);
-                        } else {
-                            self.compile_stmt(s);
-                        }
-                    } else {
-                        self.compile_stmt(s);
-                    }
-                }
-                self.code.patch_body_end(default_idx);
-            }
-            // Repeat loop (repeat while / repeat until)
-            Stmt::Loop {
-                init,
-                cond,
-                step,
-                body,
-                repeat,
-                label,
-            } if *repeat => {
-                let (pre_stmts, loop_body, post_stmts) = self.expand_loop_phasers(body);
-                if let Some(init_stmt) = init {
-                    self.compile_stmt(init_stmt);
-                }
-                for s in &pre_stmts {
-                    self.compile_stmt(s);
-                }
-                // Layout: [RepeatLoop] [body..] [cond..]
-                let loop_idx = self.code.emit(OpCode::RepeatLoop {
-                    cond_end: 0,
-                    body_end: 0,
-                    label: label.clone(),
-                });
-                // Compile body
-                for s in &loop_body {
-                    self.compile_stmt(s);
-                }
-                self.code.patch_repeat_cond_end(loop_idx);
-                // Compile condition (or push True if none)
-                if let Some(cond_expr) = cond {
-                    self.compile_expr(cond_expr);
-                } else {
-                    self.code.emit(OpCode::LoadTrue);
-                }
-                // Compile step (if any)
-                if let Some(step_expr) = step {
-                    self.compile_expr(step_expr);
-                    self.code.emit(OpCode::Pop);
-                }
-                self.code.patch_loop_end(loop_idx);
-                for s in &post_stmts {
-                    self.compile_stmt(s);
-                }
-            }
-            Stmt::Loop { .. } => unreachable!("loop repeat flag is exhaustive"),
-            // --- No-ops: these statements are handled elsewhere ---
-            // CATCH/CONTROL are handled by try expressions, not standalone
-            Stmt::Catch(_) | Stmt::Control(_) => {}
-            // HasDecl/MethodDecl/DoesDecl outside class context are no-ops
-            Stmt::HasDecl { .. } | Stmt::MethodDecl { .. } | Stmt::DoesDecl { .. } => {}
-
-            // --- Take (gather/take) ---
-            Stmt::Take(expr) => {
-                self.compile_expr(expr);
-                self.code.emit(OpCode::Take);
-            }
-
-            // --- React: just run the body block ---
-            Stmt::React { body } => {
-                for s in body {
-                    self.compile_stmt(s);
-                }
-            }
-
-            // --- Package scope ---
-            Stmt::Package { name, body } => {
-                if body.is_empty() {
-                    // unit module/package — set package for the rest of the scope
-                    self.current_package = name.clone();
-                } else {
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    let pkg_idx = self.code.emit(OpCode::PackageScope {
-                        name_idx,
-                        body_end: 0,
-                    });
-                    let saved_package = self.current_package.clone();
-                    self.current_package = name.clone();
-                    for s in body {
-                        self.compile_stmt(s);
-                    }
-                    self.current_package = saved_package;
-                    self.code.patch_body_end(pkg_idx);
-                }
-            }
-
-            // --- Phaser (BEGIN/END) ---
-            Stmt::Phaser {
-                kind: PhaserKind::Begin | PhaserKind::Check | PhaserKind::Init,
-                body,
-            } => {
-                // BEGIN: compile body inline (runs immediately)
-                for s in body {
-                    self.compile_stmt(s);
-                }
-            }
-            Stmt::Phaser {
-                kind: PhaserKind::End,
-                body,
-            } => {
-                // END: store body in stmt pool for deferred execution
-                let end_stmt = Stmt::Phaser {
-                    kind: PhaserKind::End,
-                    body: body.clone(),
-                };
-                let idx = self.code.add_stmt(end_stmt);
-                self.code.emit(OpCode::PhaserEnd(idx));
-            }
-            Stmt::Phaser { .. } => {}
-
-            // --- SubDecl: delegate to interpreter AND compile body ---
-            Stmt::SubDecl {
-                name,
-                params,
-                param_defs,
-                body,
-                multi,
-                ..
-            } => {
-                let idx = self.code.add_stmt(stmt.clone());
-                self.code.emit(OpCode::RegisterSub(idx));
-                // Also compile the body to bytecode for VM-native dispatch
-                self.compile_sub_body(name, params, param_defs, body, *multi);
-            }
-            Stmt::TokenDecl { .. } | Stmt::RuleDecl { .. } => {
-                let idx = self.code.add_stmt(stmt.clone());
-                self.code.emit(OpCode::RegisterToken(idx));
-            }
-            Stmt::ProtoDecl {
-                params, param_defs, ..
-            } => {
-                let _ = (params.len(), param_defs.len());
-                let idx = self.code.add_stmt(stmt.clone());
-                self.code.emit(OpCode::RegisterProtoSub(idx));
-            }
-            Stmt::ProtoToken { .. } => {
-                let idx = self.code.add_stmt(stmt.clone());
-                self.code.emit(OpCode::RegisterProtoToken(idx));
-            }
-            Stmt::Use { module, arg } if module == "lib" && arg.is_some() => {
-                if let Some(expr) = arg {
-                    self.compile_expr(expr);
-                    self.code.emit(OpCode::UseLibPath);
-                }
-            }
-            Stmt::Use { module, arg } if module == "lib" && arg.is_none() => {}
-            Stmt::Use { module, .. }
-                if module == "v6"
-                    || module == "customtrait"
-                    || module == "isms"
-                    || module == "MONKEY-TYPING"
-                    || module == "nqp" => {}
-            Stmt::Use { module, .. } if module == "Test" || module.starts_with("Test::") => {
-                let name_idx = self.code.add_constant(Value::Str(module.clone()));
-                self.code.emit(OpCode::UseModule(name_idx));
-            }
-            Stmt::Use { module, .. } => {
-                let name_idx = self.code.add_constant(Value::Str(module.clone()));
-                self.code.emit(OpCode::UseModule(name_idx));
-            }
-            Stmt::EnumDecl { .. } => {
-                let idx = self.code.add_stmt(stmt.clone());
-                self.code.emit(OpCode::RegisterEnum(idx));
-            }
-            Stmt::ClassDecl { .. } => {
-                let idx = self.code.add_stmt(stmt.clone());
-                self.code.emit(OpCode::RegisterClass(idx));
-            }
-            Stmt::RoleDecl { .. } => {
-                let idx = self.code.add_stmt(stmt.clone());
-                self.code.emit(OpCode::RegisterRole(idx));
-            }
-            Stmt::SubsetDecl { .. } => {
-                let idx = self.code.add_stmt(stmt.clone());
-                self.code.emit(OpCode::RegisterSubset(idx));
-            }
-            Stmt::Subtest { name, body } => {
-                self.compile_expr(name);
-                let idx = self.code.emit(OpCode::SubtestScope { body_end: 0 });
-                for s in body {
-                    self.compile_stmt(s);
-                }
-                self.code.patch_body_end(idx);
-            }
-            Stmt::Whenever {
-                supply,
-                param,
-                body,
-            } => {
-                self.compile_expr(supply);
-                let body_idx = self.code.add_stmt(Stmt::Block(body.clone()));
-                let param_idx = param
-                    .as_ref()
-                    .map(|p| self.code.add_constant(Value::Str(p.clone())));
-                let target_var_idx = if let Expr::Var(name) = supply {
-                    Some(self.code.add_constant(Value::Str(name.clone())))
-                } else {
-                    None
-                };
-                self.code.emit(OpCode::WheneverScope {
-                    body_idx,
-                    param_idx,
-                    target_var_idx,
-                });
-            }
-            Stmt::Let { name, index, value } => {
-                // Emit LetSave: saves current value of the variable
-                let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                let has_index = index.is_some();
-                if let Some(idx_expr) = index {
-                    self.compile_expr(idx_expr);
-                }
-                self.code.emit(OpCode::LetSave {
-                    name_idx,
-                    index_mode: has_index,
-                });
-                // Compile the assignment if value is provided
-                if let Some(val_expr) = value {
-                    if has_index {
-                        // For array index assignment: compile as Stmt::Expr(IndexAssign)
-                        let target_expr = if let Some(stripped) = name.strip_prefix('@') {
-                            Expr::ArrayVar(stripped.to_string())
-                        } else if let Some(stripped) = name.strip_prefix('%') {
-                            Expr::Var(stripped.to_string())
-                        } else {
-                            Expr::Var(name.to_string())
-                        };
-                        let assign_expr = Expr::IndexAssign {
-                            target: Box::new(target_expr),
-                            index: Box::new(index.as_ref().unwrap().as_ref().clone()),
-                            value: Box::new(val_expr.as_ref().clone()),
-                        };
-                        self.compile_expr(&assign_expr);
-                        self.code.emit(OpCode::Pop);
-                    } else {
-                        self.compile_expr(val_expr);
-                        if let Some(&slot) = self.local_map.get(name.as_str()) {
-                            self.code.emit(OpCode::SetLocal(slot));
-                        } else {
-                            let set_idx = self.code.add_constant(Value::Str(name.clone()));
-                            self.code.emit(OpCode::SetGlobal(set_idx));
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    fn compile_expr(&mut self, expr: &Expr) {
-        match expr {
-            Expr::Whatever => {
-                let idx = self.code.add_constant(Value::Num(f64::INFINITY));
-                self.code.emit(OpCode::LoadConst(idx));
-            }
-            Expr::Literal(v) => match v {
-                Value::Nil => {
-                    self.code.emit(OpCode::LoadNil);
-                }
-                Value::Bool(true) => {
-                    self.code.emit(OpCode::LoadTrue);
-                }
-                Value::Bool(false) => {
-                    self.code.emit(OpCode::LoadFalse);
-                }
-                _ => {
-                    let idx = self.code.add_constant(v.clone());
-                    self.code.emit(OpCode::LoadConst(idx));
-                }
-            },
-            Expr::Var(name) => {
-                // Compile-time package/module variables
-                if name == "?PACKAGE" || name == "?MODULE" {
-                    let idx = self
-                        .code
-                        .add_constant(Value::Package(self.current_package.clone()));
-                    self.code.emit(OpCode::LoadConst(idx));
-                } else if let Some(&slot) = self.local_map.get(name.as_str()) {
-                    self.code.emit(OpCode::GetLocal(slot));
-                } else {
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    self.code.emit(OpCode::GetGlobal(name_idx));
-                }
-            }
-            Expr::ArrayVar(name) => {
-                let name_idx = self.code.add_constant(Value::Str(format!("@{}", name)));
-                self.code.emit(OpCode::GetArrayVar(name_idx));
-            }
-            Expr::HashVar(name) => {
-                let name_idx = self.code.add_constant(Value::Str(format!("%{}", name)));
-                self.code.emit(OpCode::GetHashVar(name_idx));
-            }
-            Expr::BareWord(name) => {
-                // Check if this bare word is a local variable (e.g., from constant declaration)
-                if let Some(&slot) = self.local_map.get(name.as_str()) {
-                    self.code.emit(OpCode::GetLocal(slot));
-                } else {
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    self.code.emit(OpCode::GetBareWord(name_idx));
-                }
-            }
-            Expr::Unary { op, expr } => match op {
-                TokenKind::Minus => {
-                    self.compile_expr(expr);
-                    self.code.emit(OpCode::Negate);
-                }
-                TokenKind::Bang => {
-                    self.compile_expr(expr);
-                    self.code.emit(OpCode::Not);
-                }
-                TokenKind::Question => {
-                    self.compile_expr(expr);
-                    self.code.emit(OpCode::BoolCoerce);
-                }
-                TokenKind::Plus => {
-                    self.compile_expr(expr);
-                    self.code.emit(OpCode::NumCoerce);
-                }
-                TokenKind::Tilde => {
-                    self.compile_expr(expr);
-                    self.code.emit(OpCode::StrCoerce);
-                }
-                TokenKind::Caret => {
-                    self.compile_expr(expr);
-                    self.code.emit(OpCode::UptoRange);
-                }
-                TokenKind::Ident(name) if name == "so" => {
-                    self.compile_expr(expr);
-                    self.code.emit(OpCode::BoolCoerce);
-                }
-                TokenKind::PlusPlus => {
-                    if let Expr::Var(name) = expr.as_ref() {
-                        let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                        self.code.emit(OpCode::PreIncrement(name_idx));
-                    } else if let Some(var_name) = Self::extract_vardecl_name(expr) {
-                        // ++state $ — compile the VarDecl, then increment
-                        self.compile_expr(expr);
-                        self.code.emit(OpCode::Pop);
-                        let name_idx = self.code.add_constant(Value::Str(var_name));
-                        self.code.emit(OpCode::PreIncrement(name_idx));
-                    } else {
-                        self.code.emit(OpCode::LoadNil);
-                    }
-                }
-                TokenKind::MinusMinus => {
-                    if let Expr::Var(name) = expr.as_ref() {
-                        let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                        self.code.emit(OpCode::PreDecrement(name_idx));
-                    } else if let Some(var_name) = Self::extract_vardecl_name(expr) {
-                        self.compile_expr(expr);
-                        self.code.emit(OpCode::Pop);
-                        let name_idx = self.code.add_constant(Value::Str(var_name));
-                        self.code.emit(OpCode::PreDecrement(name_idx));
-                    } else {
-                        self.code.emit(OpCode::LoadNil);
-                    }
-                }
-                TokenKind::IntBitNeg => {
-                    self.compile_expr(expr);
-                    self.code.emit(OpCode::IntBitNeg);
-                }
-                TokenKind::BoolBitNeg => {
-                    self.compile_expr(expr);
-                    self.code.emit(OpCode::BoolBitNeg);
-                }
-                TokenKind::Pipe => {
-                    self.compile_expr(expr);
-                    self.code.emit(OpCode::MakeSlip);
-                }
-                _ => {
-                    // Fallback: delegate to interpreter for unsupported prefix operators
-                    let prefix_expr = Expr::Unary {
-                        op: op.clone(),
-                        expr: expr.clone(),
-                    };
-                    let idx = self.code.add_stmt(Stmt::Expr(prefix_expr));
-                    self.code.emit(OpCode::EvalAstExpr(idx));
-                }
-            },
-            Expr::Binary { left, op, right } => {
-                // Short-circuit operators
-                match op {
-                    TokenKind::AndAnd => {
-                        self.compile_expr(left);
-                        self.code.emit(OpCode::Dup);
-                        let jump_end = self.code.emit(OpCode::JumpIfFalse(0));
-                        self.code.emit(OpCode::Pop);
-                        self.compile_expr(right);
-                        self.code.patch_jump(jump_end);
-                        return;
-                    }
-                    TokenKind::OrOr | TokenKind::OrWord => {
-                        self.compile_expr(left);
-                        self.code.emit(OpCode::Dup);
-                        let jump_end = self.code.emit(OpCode::JumpIfTrue(0));
-                        self.code.emit(OpCode::Pop);
-                        self.compile_expr(right);
-                        self.code.patch_jump(jump_end);
-                        return;
-                    }
-                    TokenKind::SlashSlash | TokenKind::OrElse => {
-                        self.compile_expr(left);
-                        self.code.emit(OpCode::Dup);
-                        let jump_end = self.code.emit(OpCode::JumpIfNotNil(0));
-                        self.code.emit(OpCode::Pop);
-                        self.compile_expr(right);
-                        self.code.patch_jump(jump_end);
-                        return;
-                    }
-                    TokenKind::AndThen => {
-                        self.compile_expr(left);
-                        self.code.emit(OpCode::Dup);
-                        let jump_nil = self.code.emit(OpCode::JumpIfNil(0));
-                        self.code.emit(OpCode::Pop);
-                        self.compile_expr(right);
-                        self.code.patch_jump(jump_nil);
-                        return;
-                    }
-                    TokenKind::NotAndThen => {
-                        self.compile_expr(left);
-                        let jump_eval_right = self.code.emit(OpCode::JumpIfNil(0));
-                        self.code.emit(OpCode::Pop);
-                        self.code.emit(OpCode::LoadNil);
-                        let jump_end = self.code.emit(OpCode::Jump(0));
-                        self.code.patch_jump(jump_eval_right);
-                        self.code.emit(OpCode::Pop);
-                        self.compile_expr(right);
-                        self.code.patch_jump(jump_end);
-                        return;
-                    }
-                    TokenKind::SmartMatch | TokenKind::BangTilde => {
-                        let lhs_var = match left.as_ref() {
-                            Expr::Var(name) => Some(name.clone()),
-                            _ => None,
-                        };
-                        self.compile_expr(left);
-                        let sm_idx = self.code.emit(OpCode::SmartMatchExpr {
-                            rhs_end: 0,
-                            negate: matches!(op, TokenKind::BangTilde),
-                            lhs_var,
-                        });
-                        self.compile_expr(right);
-                        self.code.patch_smart_match_rhs_end(sm_idx);
-                        return;
-                    }
-                    _ => {}
-                }
-
-                if let Some(opcode) = Self::binary_opcode(op) {
-                    self.compile_expr(left);
-                    self.compile_expr(right);
-                    self.code.emit(opcode);
-                } else {
-                    // Fallback: delegate to interpreter for unsupported operators
-                    let expr = Expr::Binary {
-                        left: left.clone(),
-                        op: op.clone(),
-                        right: right.clone(),
-                    };
-                    let idx = self.code.add_stmt(Stmt::Expr(expr));
-                    self.code.emit(OpCode::EvalAstExpr(idx));
-                }
-            }
-            Expr::Ternary {
-                cond,
-                then_expr,
-                else_expr,
-            } => {
-                self.compile_expr(cond);
-                let jump_else = self.code.emit(OpCode::JumpIfFalse(0));
-                self.compile_expr(then_expr);
-                let jump_end = self.code.emit(OpCode::Jump(0));
-                self.code.patch_jump(jump_else);
-                self.compile_expr(else_expr);
-                self.code.patch_jump(jump_end);
-            }
-            Expr::ArrayLiteral(elems) => {
-                for elem in elems {
-                    self.compile_expr(elem);
-                }
-                self.code.emit(OpCode::MakeArray(elems.len() as u32));
-            }
-            // Expression-level function call
-            Expr::Call { name, args } => {
-                // Rewrite undefine($var) → $var = Nil (assign Nil to the variable)
-                if name == "undefine" && args.len() == 1 {
-                    let var_name = match &args[0] {
-                        Expr::Var(n) => Some(n.clone()),
-                        Expr::ArrayVar(n) => Some(format!("@{}", n)),
-                        Expr::HashVar(n) => Some(format!("%{}", n)),
-                        Expr::CodeVar(n) => Some(format!("&{}", n)),
-                        _ => None,
-                    };
-                    if let Some(vname) = var_name {
-                        // Push Nil and assign to the variable
-                        self.code.emit(OpCode::LoadNil);
-                        let name_idx = self.code.add_constant(Value::Str(vname));
-                        self.code.emit(OpCode::AssignExpr(name_idx));
-                    } else {
-                        self.code.emit(OpCode::LoadNil);
-                    }
-                }
-                // Rewrite shift(@arr)/pop(@arr) → @arr.shift()/@arr.pop() for mutability
-                else if matches!(name.as_str(), "shift" | "pop")
-                    && args.len() == 1
-                    && matches!(args[0], Expr::ArrayVar(_) | Expr::Var(_))
-                {
-                    let method_call = Expr::MethodCall {
-                        target: Box::new(args[0].clone()),
-                        name: name.clone(),
-                        args: Vec::new(),
-                        modifier: None,
-                    };
-                    self.compile_expr(&method_call);
-                }
-                // Rewrite push(@arr, val...)/unshift(@arr, val...)/append/prepend → @arr.push(val...)
-                else if matches!(name.as_str(), "push" | "unshift" | "append" | "prepend")
-                    && args.len() >= 2
-                    && matches!(args[0], Expr::ArrayVar(_) | Expr::Var(_))
-                {
-                    let method_call = Expr::MethodCall {
-                        target: Box::new(args[0].clone()),
-                        name: name.clone(),
-                        args: args[1..].to_vec(),
-                        modifier: None,
-                    };
-                    self.compile_expr(&method_call);
-                }
-                // Rewrite indir($path, { ... }) body into a callable block value so
-                // call_function("indir", ...) can execute it after switching $*CWD.
-                else if name == "indir" && args.len() >= 2 {
-                    let mut rewritten_args = args.clone();
-                    if let Expr::Block(body) = &rewritten_args[1] {
-                        rewritten_args[1] = make_anon_sub(body.clone());
-                    }
-                    let arity = rewritten_args.len() as u32;
-                    for arg in &rewritten_args {
-                        self.compile_expr(arg);
-                    }
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    self.code.emit(OpCode::CallFunc { name_idx, arity });
-                } else {
-                    // Rewrite VAR($var)/VAR(@var)/VAR(%var)/VAR(&var) → $var.VAR, etc.
-                    if name == "VAR"
-                        && args.len() == 1
-                        && matches!(
-                            args[0],
-                            Expr::Var(_) | Expr::ArrayVar(_) | Expr::HashVar(_) | Expr::CodeVar(_)
-                        )
-                    {
-                        let method_call = Expr::MethodCall {
-                            target: Box::new(args[0].clone()),
-                            name: "VAR".to_string(),
-                            args: Vec::new(),
-                            modifier: None,
-                        };
-                        self.compile_expr(&method_call);
-                        return;
-                    }
-                    let arity = args.len() as u32;
-                    for arg in args {
-                        self.compile_expr(arg);
-                    }
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    self.code.emit(OpCode::CallFunc { name_idx, arity });
-                }
-            }
-            // Method call on mutable variable target (needs writeback)
-            Expr::MethodCall {
-                target,
-                name,
-                args,
-                modifier,
-            } if matches!(
-                target.as_ref(),
-                Expr::Var(_) | Expr::ArrayVar(_) | Expr::HashVar(_) | Expr::CodeVar(_)
-            ) || Self::is_dostmt_vardecl(target) =>
-            {
-                let target_name = match target.as_ref() {
-                    Expr::Var(n) => n.clone(),
-                    Expr::ArrayVar(n) => format!("@{}", n),
-                    Expr::HashVar(n) => format!("%{}", n),
-                    Expr::CodeVar(n) => format!("&{}", n),
-                    Expr::DoStmt(stmt) => {
-                        if let Stmt::VarDecl { name, .. } = stmt.as_ref() {
-                            name.clone()
-                        } else {
-                            unreachable!()
-                        }
-                    }
-                    _ => unreachable!(),
-                };
-                self.compile_expr(target);
-                let arity = args.len() as u32;
-                for arg in args {
-                    self.compile_method_arg(arg);
-                }
-                let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                let target_name_idx = self.code.add_constant(Value::Str(target_name));
-                let modifier_idx =
-                    modifier.map(|m| self.code.add_constant(Value::Str(m.to_string())));
-                self.code.emit(OpCode::CallMethodMut {
-                    name_idx,
-                    arity,
-                    target_name_idx,
-                    modifier_idx,
-                });
-            }
-            // Method call on non-variable target (no writeback needed)
-            Expr::MethodCall {
-                target,
-                name,
-                args,
-                modifier,
-            } => {
-                self.compile_expr(target);
-                let arity = args.len() as u32;
-                for arg in args {
-                    self.compile_method_arg(arg);
-                }
-                let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                let modifier_idx =
-                    modifier.map(|m| self.code.add_constant(Value::Str(m.to_string())));
-                self.code.emit(OpCode::CallMethod {
-                    name_idx,
-                    arity,
-                    modifier_idx,
-                });
-            }
-            // Hyper method call: target».method(args)
-            Expr::HyperMethodCall { target, name, args } => {
-                self.compile_expr(target);
-                let arity = args.len() as u32;
-                for arg in args {
-                    self.compile_method_arg(arg);
-                }
-                let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                self.code.emit(OpCode::HyperMethodCall { name_idx, arity });
-            }
-            // Indexing
-            Expr::Index { target, index } => {
-                // Special case: %*ENV<key> compiles to GetEnvIndex
-                if let Expr::HashVar(name) = target.as_ref() {
-                    if name == "*ENV" {
-                        if let Expr::Literal(Value::Str(key)) = index.as_ref() {
-                            let key_idx = self.code.add_constant(Value::Str(key.clone()));
-                            self.code.emit(OpCode::GetEnvIndex(key_idx));
-                        } else {
-                            // Dynamic key - compile as runtime hash + index
-                            self.compile_expr(target);
-                            self.compile_expr(index);
-                            self.code.emit(OpCode::Index);
-                        }
-                    } else {
-                        self.compile_expr(target);
-                        self.compile_expr(index);
-                        self.code.emit(OpCode::Index);
-                    }
-                } else {
-                    self.compile_expr(target);
-                    self.compile_expr(index);
-                    self.code.emit(OpCode::Index);
-                }
-            }
-            // String interpolation
-            Expr::StringInterpolation(parts) => {
-                let n = parts.len() as u32;
-                for part in parts {
-                    self.compile_expr(part);
-                }
-                self.code.emit(OpCode::StringConcat(n));
-            }
-            // Postfix ++ on variable
-            Expr::PostfixOp {
-                op: TokenKind::PlusPlus,
-                expr,
-            } => {
-                if let Expr::Var(name) = expr.as_ref() {
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    self.code.emit(OpCode::PostIncrement(name_idx));
-                } else if let Expr::Index { target, index } = expr.as_ref() {
-                    if let Some(name) = Self::postfix_index_name(target) {
-                        self.compile_expr(index);
-                        let name_idx = self.code.add_constant(Value::Str(name));
-                        self.code.emit(OpCode::PostIncrementIndex(name_idx));
-                    } else {
-                        self.code.emit(OpCode::LoadNil);
-                    }
-                } else {
-                    self.code.emit(OpCode::LoadNil);
-                }
-            }
-            // Postfix -- on variable
-            Expr::PostfixOp {
-                op: TokenKind::MinusMinus,
-                expr,
-            } => {
-                if let Expr::Var(name) = expr.as_ref() {
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    self.code.emit(OpCode::PostDecrement(name_idx));
-                } else if let Expr::Index { target, index } = expr.as_ref() {
-                    if let Some(name) = Self::postfix_index_name(target) {
-                        self.compile_expr(index);
-                        let name_idx = self.code.add_constant(Value::Str(name));
-                        self.code.emit(OpCode::PostDecrementIndex(name_idx));
-                    } else {
-                        self.code.emit(OpCode::LoadNil);
-                    }
-                } else {
-                    self.code.emit(OpCode::LoadNil);
-                }
-            }
-            // Assignment as expression
-            Expr::AssignExpr { name, expr } => {
-                self.compile_expr(expr);
-                if let Some(&slot) = self.local_map.get(name.as_str()) {
-                    self.code.emit(OpCode::AssignExprLocal(slot));
-                } else {
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    self.code.emit(OpCode::AssignExpr(name_idx));
-                }
-            }
-            // Capture variable ($0, $1, etc.)
-            Expr::CaptureVar(name) => {
-                let name_idx = self.code.add_constant(Value::Str(format!("<{}>", name)));
-                self.code.emit(OpCode::GetCaptureVar(name_idx));
-            }
-            // Code variable (&foo)
-            Expr::CodeVar(name) => {
-                let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                self.code.emit(OpCode::GetCodeVar(name_idx));
-            }
-            // Hash literal
-            Expr::Hash(pairs) => {
-                let n = pairs.len() as u32;
-                for (key, val_opt) in pairs {
-                    // Push key as string constant
-                    let key_idx = self.code.add_constant(Value::Str(key.clone()));
-                    self.code.emit(OpCode::LoadConst(key_idx));
-                    // Push value (or Nil if none)
-                    if let Some(val_expr) = val_opt {
-                        self.compile_expr(val_expr);
-                    } else {
-                        self.code.emit(OpCode::LoadNil);
-                    }
-                }
-                self.code.emit(OpCode::MakeHash(n));
-            }
-            // Environment variable access (%*ENV<key>)
-            Expr::EnvIndex(key) => {
-                let key_idx = self.code.add_constant(Value::Str(key.clone()));
-                self.code.emit(OpCode::GetEnvIndex(key_idx));
-            }
-            // Exists check (:exists)
-            Expr::Exists(inner) => match inner.as_ref() {
-                Expr::EnvIndex(key) => {
-                    let key_idx = self.code.add_constant(Value::Str(key.clone()));
-                    self.code.emit(OpCode::ExistsEnvIndex(key_idx));
-                }
-                Expr::Index { target, index } if matches!(target.as_ref(), Expr::HashVar(name) if name == "*ENV") => {
-                    if let Expr::Literal(Value::Str(key)) = index.as_ref() {
-                        let key_idx = self.code.add_constant(Value::Str(key.clone()));
-                        self.code.emit(OpCode::ExistsEnvIndex(key_idx));
-                    } else {
-                        self.compile_expr(inner);
-                        self.code.emit(OpCode::ExistsExpr);
-                    }
-                }
-                _ => {
-                    self.compile_expr(inner);
-                    self.code.emit(OpCode::ExistsExpr);
-                }
-            },
-            // Reduction ([+] @arr)
-            Expr::Reduction { op, expr } => {
-                self.compile_expr(expr);
-                let op_idx = self.code.add_constant(Value::Str(op.clone()));
-                self.code.emit(OpCode::Reduction(op_idx));
-            }
-            // __ROUTINE__ magic
-            Expr::RoutineMagic => {
-                self.code.emit(OpCode::RoutineMagic);
-            }
-            // __BLOCK__ magic
-            Expr::BlockMagic => {
-                self.code.emit(OpCode::BlockMagic);
-            }
-            // s/// substitution
-            Expr::Subst {
-                pattern,
-                replacement,
-            } => {
-                let pattern_idx = self.code.add_constant(Value::Str(pattern.clone()));
-                let replacement_idx = self.code.add_constant(Value::Str(replacement.clone()));
-                self.code.emit(OpCode::Subst {
-                    pattern_idx,
-                    replacement_idx,
-                });
-            }
-            // S/// non-destructive substitution
-            Expr::NonDestructiveSubst {
-                pattern,
-                replacement,
-            } => {
-                let pattern_idx = self.code.add_constant(Value::Str(pattern.clone()));
-                let replacement_idx = self.code.add_constant(Value::Str(replacement.clone()));
-                self.code.emit(OpCode::NonDestructiveSubst {
-                    pattern_idx,
-                    replacement_idx,
-                });
-            }
-            // tr/// transliteration
-            Expr::Transliterate { from, to } => {
-                let from_idx = self.code.add_constant(Value::Str(from.clone()));
-                let to_idx = self.code.add_constant(Value::Str(to.clone()));
-                self.code.emit(OpCode::Transliterate { from_idx, to_idx });
-            }
-            // HyperOp (>>op<<): compile sub-expressions, delegate operation
-            Expr::HyperOp {
-                op,
-                left,
-                right,
-                dwim_left,
-                dwim_right,
-            } => {
-                self.compile_expr(left);
-                self.compile_expr(right);
-                let op_idx = self.code.add_constant(Value::Str(op.clone()));
-                self.code.emit(OpCode::HyperOp {
-                    op_idx,
-                    dwim_left: *dwim_left,
-                    dwim_right: *dwim_right,
-                });
-            }
-            // MetaOp (Rop, Xop, Zop): compile sub-expressions, delegate operation
-            Expr::MetaOp {
-                meta,
-                op,
-                left,
-                right,
-            } => {
-                self.compile_expr(left);
-                self.compile_expr(right);
-                let meta_idx = self.code.add_constant(Value::Str(meta.clone()));
-                let op_idx = self.code.add_constant(Value::Str(op.clone()));
-                self.code.emit(OpCode::MetaOp { meta_idx, op_idx });
-            }
-            // InfixFunc (atan2, sprintf): compile sub-expressions, delegate operation
-            Expr::InfixFunc {
-                name,
-                left,
-                right,
-                modifier,
-            } => {
-                self.compile_expr(left);
-                for r in right {
-                    self.compile_expr(r);
-                }
-                let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                let modifier_idx = modifier
-                    .as_ref()
-                    .map(|m| self.code.add_constant(Value::Str(m.clone())));
-                self.code.emit(OpCode::InfixFunc {
-                    name_idx,
-                    right_arity: right.len() as u32,
-                    modifier_idx,
-                });
-            }
-            Expr::Try { body, catch } => {
-                self.compile_try(body, catch);
-            }
-            Expr::DoBlock { .. } => {
-                if let Expr::DoBlock { body, label } = expr {
-                    self.compile_do_block_expr(body, label);
-                }
-            }
-            Expr::DoStmt(stmt) => match stmt.as_ref() {
-                Stmt::If {
-                    cond,
-                    then_branch,
-                    else_branch,
-                } if Self::do_if_branch_supported(then_branch)
-                    && Self::do_if_branch_supported(else_branch) =>
-                {
-                    self.compile_do_if_expr(cond, then_branch, else_branch);
-                }
-                Stmt::Given { topic, body } => {
-                    self.compile_expr(topic);
-                    let given_idx = self.code.emit(OpCode::DoGivenExpr { body_end: 0 });
-                    for s in body {
-                        self.compile_stmt(s);
-                    }
-                    self.code.patch_body_end(given_idx);
-                }
-                Stmt::Assign { name, expr, .. } => {
-                    // do $var = expr → returns the assigned value
-                    self.compile_expr(expr);
-                    // Duplicate the value so we can assign AND return it
-                    self.code.emit(OpCode::Dup);
-                    if let Some(&slot) = self.local_map.get(name.as_str()) {
-                        self.code.emit(OpCode::SetLocal(slot));
-                    } else {
-                        let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                        self.code.emit(OpCode::SetGlobal(name_idx));
-                    }
-                }
-                Stmt::VarDecl {
-                    name,
-                    expr,
-                    is_state,
-                    ..
-                } => {
-                    // my $x = expr in expression context → declare, assign, return value
-                    self.compile_expr(expr);
-                    if *is_state {
-                        let slot = self.alloc_local(name);
-                        let key =
-                            format!("__state_{}", STATE_COUNTER.fetch_add(1, Ordering::Relaxed));
-                        let key_idx = self.code.add_constant(Value::Str(key.clone()));
-                        self.code.state_locals.push((slot as usize, key.clone()));
-                        self.code.emit(OpCode::StateVarInit(slot, key_idx));
-                        self.code.emit(OpCode::GetLocal(slot));
-                    } else {
-                        self.code.emit(OpCode::Dup);
-                        if let Some(&slot) = self.local_map.get(name.as_str()) {
-                            self.code.emit(OpCode::SetLocal(slot));
-                        } else {
-                            let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                            self.code.emit(OpCode::SetGlobal(name_idx));
-                        }
-                    }
-                }
-                Stmt::Expr(inner_expr) => {
-                    self.compile_expr(inner_expr);
-                }
-                Stmt::For {
-                    iterable,
-                    param,
-                    params,
-                    body,
-                    label,
-                } => {
-                    self.compile_do_for_expr(iterable, param, params, body, label);
-                }
-                Stmt::ClassDecl { name, .. } => {
-                    // Register the class and return the type object
-                    self.compile_stmt(stmt);
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    self.code.emit(OpCode::GetBareWord(name_idx));
-                }
-                _ => {
-                    self.compile_stmt(stmt);
-                    self.code.emit(OpCode::LoadNil);
-                }
-            },
-            Expr::Gather(_) => {
-                if let Expr::Gather(body) = expr {
-                    let idx = self.code.add_stmt(Stmt::Block(body.clone()));
-                    self.code.emit(OpCode::MakeGather(idx));
-                }
-            }
-            Expr::CallOn { target, args } => {
-                if let Expr::CodeVar(name) = target.as_ref() {
-                    for arg in args {
-                        self.compile_expr(arg);
-                    }
-                    let name_idx = self.code.add_constant(Value::Str(name.clone()));
-                    self.code.emit(OpCode::CallOnCodeVar {
-                        name_idx,
-                        arity: args.len() as u32,
-                    });
-                } else {
-                    self.compile_expr(target);
-                    for arg in args {
-                        self.compile_expr(arg);
-                    }
-                    self.code.emit(OpCode::CallOnValue {
-                        arity: args.len() as u32,
-                    });
-                }
-            }
-            Expr::AnonSub(body) => {
-                let idx = self.code.add_stmt(Stmt::Block(body.clone()));
-                self.code.emit(OpCode::MakeAnonSub(idx));
-            }
-            Expr::AnonSubParams { params, body } => {
-                let idx = self.code.add_stmt(Stmt::SubDecl {
-                    name: String::new(),
-                    params: params.clone(),
-                    param_defs: Vec::new(),
-                    body: body.clone(),
-                    multi: false,
-                    is_export: false,
-                });
-                self.code.emit(OpCode::MakeAnonSubParams(idx));
-            }
-            Expr::Lambda { param, body } => {
-                let idx = self.code.add_stmt(Stmt::SubDecl {
-                    name: String::new(),
-                    params: if param.is_empty() {
-                        Vec::new()
-                    } else {
-                        vec![param.clone()]
-                    },
-                    param_defs: Vec::new(),
-                    body: body.clone(),
-                    multi: false,
-                    is_export: false,
-                });
-                self.code.emit(OpCode::MakeLambda(idx));
-            }
-            Expr::IndexAssign {
-                target,
-                index,
-                value,
-            } => {
-                if let Some(name) = Self::index_assign_target_name(target) {
-                    self.compile_expr(value);
-                    self.compile_expr(index);
-                    let name_idx = self.code.add_constant(Value::Str(name));
-                    self.code.emit(OpCode::IndexAssignExprNamed(name_idx));
-                } else if let Some((name, inner_index)) = Self::index_assign_nested_target(target) {
-                    self.compile_expr(value);
-                    self.compile_expr(index);
-                    self.compile_expr(inner_index);
-                    let name_idx = self.code.add_constant(Value::Str(name));
-                    self.code.emit(OpCode::IndexAssignExprNested(name_idx));
-                } else {
-                    self.code.emit(OpCode::IndexAssignInvalid);
-                }
-            }
-            Expr::IndirectTypeLookup(inner) => {
-                self.compile_expr(inner);
-                self.code.emit(OpCode::IndirectTypeLookup);
-            }
-            Expr::ControlFlow { kind, label } => match kind {
-                crate::ast::ControlFlowKind::Last => {
-                    self.code.emit(OpCode::Last(label.clone()));
-                }
-                crate::ast::ControlFlowKind::Next => {
-                    self.code.emit(OpCode::Next(label.clone()));
-                }
-                crate::ast::ControlFlowKind::Redo => {
-                    self.code.emit(OpCode::Redo(label.clone()));
-                }
-            },
-            // Block inlining: compile inline if no placeholders
-            Expr::Block(stmts) => {
-                if Self::has_block_placeholders(stmts) {
-                    let idx = self.code.add_stmt(Stmt::Block(stmts.clone()));
-                    self.code.emit(OpCode::MakeBlockClosure(idx));
-                } else {
-                    self.compile_block_inline(stmts);
-                }
-            }
-            // Remaining expression forms not yet bytecode-native.
-            Expr::PostfixOp { .. } => {
-                self.code.emit(OpCode::LoadNil);
-            }
-        }
-    }
-
-    fn binary_opcode(op: &TokenKind) -> Option<OpCode> {
-        match op {
-            TokenKind::Plus => Some(OpCode::Add),
-            TokenKind::Minus => Some(OpCode::Sub),
-            TokenKind::Star => Some(OpCode::Mul),
-            TokenKind::Slash => Some(OpCode::Div),
-            TokenKind::Percent => Some(OpCode::Mod),
-            TokenKind::StarStar => Some(OpCode::Pow),
-            TokenKind::Tilde => Some(OpCode::Concat),
-            TokenKind::EqEq => Some(OpCode::NumEq),
-            TokenKind::BangEq => Some(OpCode::NumNe),
-            TokenKind::Lt => Some(OpCode::NumLt),
-            TokenKind::Lte => Some(OpCode::NumLe),
-            TokenKind::Gt => Some(OpCode::NumGt),
-            TokenKind::Gte => Some(OpCode::NumGe),
-            TokenKind::Ident(name) if name == "eq" => Some(OpCode::StrEq),
-            TokenKind::Ident(name) if name == "ne" => Some(OpCode::StrNe),
-            TokenKind::Ident(name) if name == "lt" => Some(OpCode::StrLt),
-            TokenKind::Ident(name) if name == "gt" => Some(OpCode::StrGt),
-            TokenKind::Ident(name) if name == "before" => Some(OpCode::Before),
-            TokenKind::Ident(name) if name == "after" => Some(OpCode::After),
-            TokenKind::Ident(name) if name == "le" => Some(OpCode::StrLe),
-            TokenKind::Ident(name) if name == "ge" => Some(OpCode::StrGe),
-            TokenKind::DotDot => Some(OpCode::MakeRange),
-            TokenKind::DotDotCaret => Some(OpCode::MakeRangeExcl),
-            TokenKind::CaretDotDot => Some(OpCode::MakeRangeExclStart),
-            TokenKind::CaretDotDotCaret => Some(OpCode::MakeRangeExclBoth),
-            // Smart match (~~, !~~): handled via interpreter fallback (needs $_ binding)
-            // Three-way comparison
-            TokenKind::LtEqGt => Some(OpCode::Spaceship),
-            TokenKind::Ident(name) if name == "cmp" => Some(OpCode::Cmp),
-            TokenKind::Ident(name) if name == "leg" => Some(OpCode::Leg),
-            // Identity/value equality
-            TokenKind::EqEqEq => Some(OpCode::StrictEq),
-            TokenKind::Ident(name) if name == "eqv" => Some(OpCode::Eqv),
-            TokenKind::Ident(name) if name == "=~=" => Some(OpCode::ApproxEq),
-            TokenKind::Ident(name) if name == "=:=" => Some(OpCode::ContainerEq),
-            // Divisibility
-            TokenKind::PercentPercent => Some(OpCode::DivisibleBy),
-            TokenKind::BangPercentPercent => Some(OpCode::NotDivisibleBy),
-            // Keyword math
-            TokenKind::Ident(name) if name == "div" => Some(OpCode::IntDiv),
-            TokenKind::Ident(name) if name == "mod" => Some(OpCode::IntMod),
-            TokenKind::Ident(name) if name == "gcd" => Some(OpCode::Gcd),
-            TokenKind::Ident(name) if name == "lcm" => Some(OpCode::Lcm),
-            TokenKind::Ident(name) if name == "min" => Some(OpCode::InfixMin),
-            TokenKind::Ident(name) if name == "max" => Some(OpCode::InfixMax),
-            // Repetition
-            TokenKind::Ident(name) if name == "x" => Some(OpCode::StringRepeat),
-            TokenKind::Ident(name) if name == "but" => Some(OpCode::ButMixin),
-            TokenKind::Ident(name) if name == "isa" => Some(OpCode::Isa),
-            TokenKind::Ident(name) if name == "does" => Some(OpCode::Does),
-            TokenKind::Ident(name) if name == "xx" => Some(OpCode::ListRepeat),
-            TokenKind::Ident(name) if name == "o" => Some(OpCode::FunctionCompose),
-            // Pair
-            TokenKind::FatArrow => Some(OpCode::MakePair),
-            // Bitwise
-            TokenKind::BitAnd => Some(OpCode::BitAnd),
-            TokenKind::BitOr => Some(OpCode::BitOr),
-            TokenKind::Ident(name) if name == "?|" => Some(OpCode::BoolBitOr),
-            TokenKind::Ident(name) if name == "?&" => Some(OpCode::BoolBitAnd),
-            TokenKind::Ident(name) if name == "?^" => Some(OpCode::BoolBitXor),
-            TokenKind::BitXor => Some(OpCode::BitXor),
-            TokenKind::BitShiftLeft => Some(OpCode::BitShiftLeft),
-            TokenKind::BitShiftRight => Some(OpCode::BitShiftRight),
-            // Set operations
-            TokenKind::SetElem => Some(OpCode::SetElem),
-            TokenKind::SetCont => Some(OpCode::SetCont),
-            TokenKind::SetUnion => Some(OpCode::SetUnion),
-            TokenKind::SetIntersect => Some(OpCode::SetIntersect),
-            TokenKind::SetDiff => Some(OpCode::SetDiff),
-            TokenKind::SetSymDiff => Some(OpCode::SetSymDiff),
-            TokenKind::SetSubset => Some(OpCode::SetSubset),
-            TokenKind::SetSuperset => Some(OpCode::SetSuperset),
-            TokenKind::SetStrictSubset => Some(OpCode::SetStrictSubset),
-            TokenKind::SetStrictSuperset => Some(OpCode::SetStrictSuperset),
-            TokenKind::Pipe => Some(OpCode::JunctionAny),
-            TokenKind::Ampersand => Some(OpCode::JunctionAll),
-            TokenKind::Caret => Some(OpCode::JunctionOne),
-            TokenKind::DotDotDot => Some(OpCode::Sequence { exclude_end: false }),
-            TokenKind::DotDotDotCaret => Some(OpCode::Sequence { exclude_end: true }),
-            _ => None,
-        }
-    }
-
-    fn is_normalized_stmt_call_name(name: &str) -> bool {
-        matches!(
-            name,
-            "shift"
-                | "pop"
-                | "push"
-                | "unshift"
-                | "append"
-                | "prepend"
-                | "undefine"
-                | "VAR"
-                | "indir"
-                | "plan"
-                | "done-testing"
-                | "skip"
-                | "skip-rest"
-                | "bail-out"
-                | "ok"
-                | "nok"
-                | "is"
-                | "isnt"
-                | "diag"
-                | "pass"
-                | "flunk"
-                | "cmp-ok"
-                | "like"
-                | "unlike"
-                | "is-deeply"
-                | "is-approx"
-                | "isa-ok"
-                | "lives-ok"
-                | "dies-ok"
-                | "eval-lives-ok"
-                | "eval-dies-ok"
-                | "throws-like"
-                | "force_todo"
-                | "force-todo"
-                | "is_run"
-                | "make-temp-dir"
-                | "make-temp-file"
-        )
-    }
-
-    fn rewrite_stmt_call_args(name: &str, args: &[CallArg]) -> Vec<CallArg> {
-        let rewrites_needed = matches!(name, "lives-ok" | "dies-ok" | "throws-like" | "is_run");
-        if !rewrites_needed {
-            return args.to_vec();
-        }
-        let mut positional_index = 0usize;
-        args.iter()
-            .map(|arg| match arg {
-                CallArg::Positional(expr) => {
-                    let rewritten = if matches!(name, "lives-ok" | "dies-ok" | "throws-like")
-                        && positional_index == 0
-                    {
-                        match expr {
-                            Expr::Block(body) => make_anon_sub(body.clone()),
-                            _ => expr.clone(),
-                        }
-                    } else if name == "is_run" && positional_index == 1 {
-                        Self::rewrite_hash_block_values(expr)
-                    } else {
-                        expr.clone()
-                    };
-                    positional_index += 1;
-                    CallArg::Positional(rewritten)
-                }
-                CallArg::Named { name, value } => CallArg::Named {
-                    name: name.clone(),
-                    value: value.clone(),
-                },
-            })
-            .collect()
-    }
-
-    /// Rewrite block values inside a hash literal to anonymous subs.
-    /// Used for `is_run`'s expectation hash: `{ out => { ... } }`.
-    fn rewrite_hash_block_values(expr: &Expr) -> Expr {
-        if let Expr::Hash(pairs) = expr {
-            let rewritten_pairs = pairs
-                .iter()
-                .map(|(name, value)| {
-                    let rewritten_value = value.as_ref().map(|v| {
-                        if let Expr::Block(body) = v {
-                            make_anon_sub(body.clone())
-                        } else {
-                            v.clone()
-                        }
-                    });
-                    (name.clone(), rewritten_value)
-                })
-                .collect();
-            Expr::Hash(rewritten_pairs)
-        } else {
-            expr.clone()
-        }
-    }
-
-    fn has_phasers(stmts: &[Stmt]) -> bool {
-        stmts
-            .iter()
-            .any(|s| matches!(s, Stmt::Phaser { kind, .. } if matches!(kind, PhaserKind::Enter | PhaserKind::Leave | PhaserKind::Keep | PhaserKind::Undo | PhaserKind::First | PhaserKind::Next | PhaserKind::Last)))
-    }
-
-    /// Check if a block body contains placeholder variables ($^a, $^b, etc.)
-    fn has_block_placeholders(stmts: &[Stmt]) -> bool {
-        for stmt in stmts {
-            if Self::stmt_has_placeholder(stmt) {
-                return true;
-            }
-        }
-        false
-    }
-
-    /// Compile a method call argument. Named args (AssignExpr) are
-    /// compiled as Pair values so they survive VM execution.
-    fn compile_method_arg(&mut self, arg: &Expr) {
-        if let Expr::AssignExpr { name, expr } = arg {
-            self.compile_expr(&Expr::Literal(Value::Str(name.clone())));
-            self.compile_expr(expr);
-            self.code.emit(OpCode::MakePair);
-        } else {
-            self.compile_expr(arg);
-        }
-    }
-
-    fn stmt_has_placeholder(stmt: &Stmt) -> bool {
-        match stmt {
-            Stmt::Expr(e) | Stmt::Return(e) | Stmt::Die(e) | Stmt::Fail(e) | Stmt::Take(e) => {
-                Self::expr_has_placeholder(e)
-            }
-            Stmt::VarDecl { expr, .. } | Stmt::Assign { expr, .. } => {
-                Self::expr_has_placeholder(expr)
-            }
-            Stmt::Say(es) | Stmt::Print(es) | Stmt::Note(es) => {
-                es.iter().any(Self::expr_has_placeholder)
-            }
-            Stmt::If {
-                cond,
-                then_branch,
-                else_branch,
-            } => {
-                Self::expr_has_placeholder(cond)
-                    || then_branch.iter().any(Self::stmt_has_placeholder)
-                    || else_branch.iter().any(Self::stmt_has_placeholder)
-            }
-            Stmt::Block(stmts) => stmts.iter().any(Self::stmt_has_placeholder),
-            _ => false,
-        }
-    }
-
-    fn expr_has_placeholder(expr: &Expr) -> bool {
-        match expr {
-            Expr::Var(name) => name.starts_with("$^"),
-            Expr::Binary { left, right, .. } => {
-                Self::expr_has_placeholder(left) || Self::expr_has_placeholder(right)
-            }
-            Expr::Unary { expr, .. } => Self::expr_has_placeholder(expr),
-            Expr::Ternary {
-                cond,
-                then_expr,
-                else_expr,
-            } => {
-                Self::expr_has_placeholder(cond)
-                    || Self::expr_has_placeholder(then_expr)
-                    || Self::expr_has_placeholder(else_expr)
-            }
-            Expr::Call { args, .. } => args.iter().any(Self::expr_has_placeholder),
-            Expr::MethodCall { target, args, .. } | Expr::HyperMethodCall { target, args, .. } => {
-                Self::expr_has_placeholder(target) || args.iter().any(Self::expr_has_placeholder)
-            }
-            Expr::Index { target, index } | Expr::IndexAssign { target, index, .. } => {
-                Self::expr_has_placeholder(target) || Self::expr_has_placeholder(index)
-            }
-            Expr::StringInterpolation(parts) | Expr::ArrayLiteral(parts) => {
-                parts.iter().any(Self::expr_has_placeholder)
-            }
-            _ => false,
-        }
-    }
-
-    /// Compile a SubDecl body to a CompiledFunction and store it.
-    fn compile_sub_body(
-        &mut self,
-        name: &str,
-        params: &[String],
-        param_defs: &[crate::ast::ParamDef],
-        body: &[Stmt],
-        multi: bool,
-    ) {
-        let mut sub_compiler = Compiler::new();
-        // Pre-allocate locals for parameters
-        for param in params {
-            sub_compiler.alloc_local(param);
-        }
-        // Also allocate from param_defs in case param names differ
-        for pd in param_defs {
-            if !pd.name.is_empty() {
-                sub_compiler.alloc_local(&pd.name);
-            }
-        }
-        // Hoist sub declarations within the sub body
-        sub_compiler.hoist_sub_decls(body);
-        // Compile body statements; last Stmt::Expr should NOT emit Pop (implicit return)
-        for (i, stmt) in body.iter().enumerate() {
-            let is_last = i == body.len() - 1;
-            if is_last && let Stmt::Expr(expr) = stmt {
-                sub_compiler.compile_expr(expr);
-                // Don't emit Pop — leave value on stack as implicit return
-                continue;
-            }
-            sub_compiler.compile_stmt(stmt);
-        }
-
-        let arity = param_defs.iter().filter(|p| !p.slurpy && !p.named).count();
-        let key = if multi {
-            let type_sig: Vec<String> = param_defs
-                .iter()
-                .map(|pd| pd.type_constraint.clone().unwrap_or_default())
-                .collect();
-            format!(
-                "{}::{}{}",
-                self.current_package,
-                name,
-                if !type_sig.is_empty() {
-                    format!("/{}:{}", arity, type_sig.join(","))
-                } else {
-                    format!("/{}", arity)
-                }
-            )
-        } else {
-            // Include arity in key to avoid collisions between same-named
-            // subs with different arities in different scopes
-            format!("{}::{}/{}", self.current_package, name, arity)
-        };
-
-        let cf = CompiledFunction {
-            code: sub_compiler.code,
-            params: params.to_vec(),
-            param_defs: param_defs.to_vec(),
-        };
-        self.compiled_functions.insert(key, cf);
-    }
-
-    /// Compile Expr::Try { body, catch } to TryCatch opcode.
-    fn compile_try(&mut self, body: &[Stmt], catch: &Option<Vec<Stmt>>) {
-        // Separate CATCH/CONTROL blocks from body.
-        let mut main_stmts = Vec::new();
-        let mut catch_stmts = catch.clone();
-        let mut control_stmts: Option<Vec<Stmt>> = None;
-        for stmt in body {
-            if let Stmt::Catch(catch_body) = stmt {
-                catch_stmts = Some(catch_body.clone());
-            } else if let Stmt::Control(control_body) = stmt {
-                control_stmts = Some(control_body.clone());
-            } else {
-                main_stmts.push(stmt.clone());
-            }
-        }
-        // Emit TryCatch placeholder.
-        let try_idx = self.code.emit(OpCode::TryCatch {
-            catch_start: 0,
-            control_start: 0,
-            body_end: 0,
-        });
-        // Compile main body (last Stmt::Expr/Call leaves value on stack)
-        let mut main_leaves_value = false;
-        for (i, stmt) in main_stmts.iter().enumerate() {
-            let is_last = i == main_stmts.len() - 1;
-            if is_last {
-                if let Stmt::Expr(expr) = stmt {
-                    self.compile_expr(expr);
-                    main_leaves_value = true;
-                    continue;
-                } else if let Stmt::Call { name, args } = stmt {
-                    // Compile as expression to leave value on stack
-                    let expr_args: Vec<Expr> = args
-                        .iter()
-                        .filter_map(|a| match a {
-                            CallArg::Positional(e) => Some(e.clone()),
-                            _ => None,
-                        })
-                        .collect();
-                    self.compile_expr(&Expr::Call {
-                        name: name.clone(),
-                        args: expr_args,
-                    });
-                    main_leaves_value = true;
-                    continue;
-                }
-            }
-            self.compile_stmt(stmt);
-        }
-        if !main_leaves_value {
-            self.code.emit(OpCode::LoadNil);
-        }
-        // Jump over catch/control on success.
-        let jump_end = self.code.emit(OpCode::Jump(0));
-        // Patch catch_start.
-        self.code.patch_try_catch_start(try_idx);
-        // Compile catch block.
-        let mut jump_after_catch = None;
-        if let Some(ref catch_body) = catch_stmts {
-            for stmt in catch_body {
-                self.compile_stmt(stmt);
-            }
-            if control_stmts.is_some() {
-                jump_after_catch = Some(self.code.emit(OpCode::Jump(0)));
-            }
-        }
-        // catch result is Nil
-        self.code.emit(OpCode::LoadNil);
-        // Patch control_start.
-        self.code.patch_try_control_start(try_idx);
-        // Compile control block.
-        if let Some(ref control_body) = control_stmts {
-            for stmt in control_body {
-                self.compile_stmt(stmt);
-            }
-            // control result is Nil
-            self.code.emit(OpCode::LoadNil);
-        }
-        // Patch body_end and jump targets.
-        self.code.patch_try_body_end(try_idx);
-        self.code.patch_jump(jump_end);
-        if let Some(j) = jump_after_catch {
-            self.code.patch_jump(j);
-        }
-    }
-
-    fn compile_do_block_expr(&mut self, body: &[Stmt], label: &Option<String>) {
-        let idx = self.code.emit(OpCode::DoBlockExpr {
-            body_end: 0,
-            label: label.clone(),
-        });
-        self.compile_block_inline(body);
-        self.code.patch_body_end(idx);
-    }
-
-    fn compile_do_if_expr(&mut self, cond: &Expr, then_branch: &[Stmt], else_branch: &[Stmt]) {
-        self.compile_expr(cond);
-        let jump_else = self.code.emit(OpCode::JumpIfFalse(0));
-        self.compile_block_inline(then_branch);
-        let jump_end = self.code.emit(OpCode::Jump(0));
-        self.code.patch_jump(jump_else);
-
-        if else_branch.is_empty() {
-            let empty_idx = self.code.add_constant(Value::Slip(vec![]));
-            self.code.emit(OpCode::LoadConst(empty_idx));
-        } else if else_branch.len() == 1 {
-            if let Stmt::If {
-                cond: inner_cond,
-                then_branch: inner_then,
-                else_branch: inner_else,
-            } = &else_branch[0]
-            {
-                self.compile_do_if_expr(inner_cond, inner_then, inner_else);
-            } else {
-                self.compile_block_inline(else_branch);
-            }
-        } else {
-            self.compile_block_inline(else_branch);
-        }
-        self.code.patch_jump(jump_end);
-    }
-
-    /// Compile `do for` expression: like a for loop but collects each iteration result.
-    fn compile_do_for_expr(
-        &mut self,
-        iterable: &Expr,
-        param: &Option<String>,
-        params: &[String],
-        body: &[Stmt],
-        label: &Option<String>,
-    ) {
-        let (_pre_stmts, mut loop_body, _post_stmts) = self.expand_loop_phasers(body);
-        let param_idx = param
-            .as_ref()
-            .map(|p| self.code.add_constant(Value::Str(p.clone())));
-        let mut bind_stmts = Vec::new();
-        if param.is_some() && param_idx.is_none() {
-            bind_stmts.push(Stmt::Assign {
-                name: param.as_ref().unwrap().clone(),
-                expr: Expr::Var("_".to_string()),
-                op: AssignOp::Assign,
-            });
-        }
-        for (i, p) in params.iter().enumerate() {
-            bind_stmts.push(Stmt::Assign {
-                name: p.clone(),
-                expr: Expr::Index {
-                    target: Box::new(Expr::Var("_".to_string())),
-                    index: Box::new(Expr::Literal(Value::Int(i as i64))),
-                },
-                op: AssignOp::Assign,
-            });
-        }
-        if !bind_stmts.is_empty() {
-            let mut merged = bind_stmts;
-            merged.extend(loop_body);
-            loop_body = merged;
-        }
-        let arity = if !params.is_empty() {
-            params.len() as u32
-        } else {
-            1
-        };
-        self.compile_expr(iterable);
-        let loop_idx = self.code.emit(OpCode::ForLoop {
-            param_idx,
-            param_local: None,
-            body_end: 0,
-            label: label.clone(),
-            arity,
-            collect: true,
-        });
-        // Compile body; last Stmt::Expr leaves value on stack (no Pop) for collection
-        for (i, s) in loop_body.iter().enumerate() {
-            let is_last = i == loop_body.len() - 1;
-            if is_last && let Stmt::Expr(expr) = s {
-                self.compile_expr(expr);
-                continue;
-            }
-            self.compile_stmt(s);
-        }
-        self.code.patch_loop_end(loop_idx);
-    }
-
-    fn do_if_branch_supported(stmts: &[Stmt]) -> bool {
-        if Self::has_phasers(stmts) {
-            return false;
-        }
-        for stmt in stmts {
-            match stmt {
-                Stmt::Given { .. } => return false,
-                Stmt::If {
-                    then_branch,
-                    else_branch,
-                    ..
-                } => {
-                    if !Self::do_if_branch_supported(then_branch)
-                        || !Self::do_if_branch_supported(else_branch)
-                    {
-                        return false;
-                    }
-                }
-                _ => {}
-            }
-        }
-        true
-    }
-
-    fn has_block_enter_leave_phasers(stmts: &[Stmt]) -> bool {
-        stmts.iter().any(|s| {
-            matches!(
-                s,
-                Stmt::Phaser {
-                    kind: PhaserKind::Enter
-                        | PhaserKind::Leave
-                        | PhaserKind::Keep
-                        | PhaserKind::Undo,
-                    ..
-                }
-            )
-        })
-    }
-
-    /// Check if a statement list contains `let` statements (not inside sub/lambda bodies).
-    fn has_let_deep(stmts: &[Stmt]) -> bool {
-        for s in stmts {
-            match s {
-                Stmt::Let { .. } => return true,
-                Stmt::Block(inner) => {
-                    if Self::has_let_deep(inner) {
-                        return true;
-                    }
-                }
-                Stmt::If {
-                    then_branch,
-                    else_branch,
-                    ..
-                } => {
-                    if Self::has_let_deep(then_branch) || Self::has_let_deep(else_branch) {
-                        return true;
-                    }
-                }
-                Stmt::Expr(expr) => {
-                    if Self::expr_has_let_deep(expr) {
-                        return true;
-                    }
-                }
-                Stmt::Call { args, .. } => {
-                    for arg in args {
-                        if let crate::ast::CallArg::Positional(expr) = arg
-                            && Self::expr_has_let_deep(expr)
-                        {
-                            return true;
-                        }
-                    }
-                }
-                _ => {}
-            }
-        }
-        false
-    }
-
-    fn expr_has_let_deep(expr: &Expr) -> bool {
-        match expr {
-            Expr::DoBlock { body, .. } => Self::has_let_deep(body),
-            Expr::Try { body, .. } => Self::has_let_deep(body),
-            Expr::Call { args, .. } => args.iter().any(Self::expr_has_let_deep),
-            Expr::MethodCall { args, target, .. } => {
-                Self::expr_has_let_deep(target) || args.iter().any(Self::expr_has_let_deep)
-            }
-            _ => false,
-        }
-    }
-
-    fn next_tmp_name(&mut self, prefix: &str) -> String {
-        let name = format!("${}{}", prefix, self.tmp_counter);
-        self.tmp_counter += 1;
-        name
-    }
-
-    fn expand_loop_phasers(&mut self, body: &[Stmt]) -> (Vec<Stmt>, Vec<Stmt>, Vec<Stmt>) {
-        if !Self::has_phasers(body) {
-            return (Vec::new(), body.to_vec(), Vec::new());
-        }
-
-        let mut enter_ph = Vec::new();
-        let mut leave_ph = Vec::new();
-        let mut first_ph = Vec::new();
-        let mut next_ph = Vec::new();
-        let mut last_ph = Vec::new();
-        let mut body_main = Vec::new();
-        for stmt in body {
-            if let Stmt::Phaser { kind, body } = stmt {
-                match kind {
-                    PhaserKind::Enter => enter_ph.push(Stmt::Block(body.clone())),
-                    PhaserKind::Leave | PhaserKind::Keep | PhaserKind::Undo => {
-                        leave_ph.push(Stmt::Block(body.clone()))
-                    }
-                    PhaserKind::First => first_ph.push(Stmt::Block(body.clone())),
-                    PhaserKind::Next => next_ph.push(Stmt::Block(body.clone())),
-                    PhaserKind::Last => last_ph.push(Stmt::Block(body.clone())),
-                    _ => body_main.push(stmt.clone()),
-                }
-            } else {
-                body_main.push(stmt.clone());
-            }
-        }
-
-        let first_var = self.next_tmp_name("__mutsu_loop_first_");
-        let ran_var = self.next_tmp_name("__mutsu_loop_ran_");
-
-        let pre = vec![
-            Stmt::VarDecl {
-                name: first_var.clone(),
-                expr: Expr::Literal(Value::Bool(true)),
-                type_constraint: None,
-                is_state: false,
-            },
-            Stmt::VarDecl {
-                name: ran_var.clone(),
-                expr: Expr::Literal(Value::Bool(false)),
-                type_constraint: None,
-                is_state: false,
-            },
-        ];
-
-        let mut loop_body = Vec::new();
-        loop_body.push(Stmt::Assign {
-            name: ran_var.clone(),
-            expr: Expr::Literal(Value::Bool(true)),
-            op: AssignOp::Assign,
-        });
-        loop_body.extend(enter_ph);
-        if !first_ph.is_empty() || !next_ph.is_empty() {
-            let mut then_branch = first_ph;
-            then_branch.push(Stmt::Assign {
-                name: first_var.clone(),
-                expr: Expr::Literal(Value::Bool(false)),
-                op: AssignOp::Assign,
-            });
-            loop_body.push(Stmt::If {
-                cond: Expr::Var(first_var.clone()),
-                then_branch,
-                else_branch: next_ph,
-            });
-        }
-        loop_body.extend(body_main);
-        loop_body.extend(leave_ph);
-
-        let post = if last_ph.is_empty() {
-            Vec::new()
-        } else {
-            vec![Stmt::If {
-                cond: Expr::Var(ran_var),
-                then_branch: last_ph,
-                else_branch: Vec::new(),
-            }]
-        };
-
-        (pre, loop_body, post)
-    }
-
-    fn is_dostmt_vardecl(expr: &Expr) -> bool {
-        matches!(expr, Expr::DoStmt(s) if matches!(s.as_ref(), Stmt::VarDecl { .. }))
-    }
-
-    /// Extract the variable name from a DoStmt(VarDecl { .. }) expression,
-    /// used for `++state $` patterns.
-    fn extract_vardecl_name(expr: &Expr) -> Option<String> {
-        if let Expr::DoStmt(stmt) = expr
-            && let Stmt::VarDecl { name, .. } = stmt.as_ref()
-        {
-            Some(name.clone())
-        } else {
-            None
-        }
-    }
-
-    fn postfix_index_name(target: &Expr) -> Option<String> {
-        match target {
-            Expr::HashVar(name) => Some(format!("%{}", name)),
-            Expr::ArrayVar(name) => Some(format!("@{}", name)),
-            Expr::Var(name) => Some(name.clone()),
-            _ => None,
-        }
-    }
-
-    fn index_assign_target_name(target: &Expr) -> Option<String> {
-        match target {
-            Expr::HashVar(name) => Some(format!("%{}", name)),
-            Expr::ArrayVar(name) => Some(format!("@{}", name)),
-            Expr::Var(name) => Some(name.clone()),
-            _ => None,
-        }
-    }
-
-    fn index_assign_nested_target(target: &Expr) -> Option<(String, &Expr)> {
-        if let Expr::Index {
-            target: inner_target,
-            index: inner_index,
-        } = target
-            && let Some(name) = Self::index_assign_target_name(inner_target)
-        {
-            return Some((name, inner_index));
-        }
-        None
-    }
-
-    /// Hoist sub declarations: emit RegisterSub for all SubDecl statements
-    /// before executing the rest of the block, so that `&name` references
-    /// are available before the sub declaration appears in source order.
-    /// Note: we only emit RegisterSub here, not compile_sub_body, because
-    /// the normal SubDecl handling will compile the body. Compiling here
-    /// would cause the compiled_functions map (which is flat) to be overwritten
-    /// by later hoists from other scopes.
-    fn hoist_sub_decls(&mut self, stmts: &[Stmt]) {
-        for stmt in stmts {
-            if let Stmt::SubDecl { .. } = stmt {
-                let idx = self.code.add_stmt(stmt.clone());
-                self.code.emit(OpCode::RegisterSub(idx));
-            }
-        }
-    }
-
-    /// Compile a block inline (for blocks without placeholders).
-    fn compile_block_inline(&mut self, stmts: &[Stmt]) {
-        if stmts.is_empty() {
-            self.code.emit(OpCode::LoadNil);
-            return;
-        }
-        // Hoist sub declarations
-        self.hoist_sub_decls(stmts);
-        for (i, stmt) in stmts.iter().enumerate() {
-            let is_last = i == stmts.len() - 1;
-            if is_last {
-                if let Stmt::Expr(expr) = stmt {
-                    self.compile_expr(expr);
-                    // Don't emit Pop — leave value on stack as block's return value
-                    return;
-                }
-                if matches!(stmt, Stmt::Given { .. }) {
-                    // given block pushes succeed value onto stack
-                    self.compile_stmt(stmt);
-                    return;
-                }
-            }
-            self.compile_stmt(stmt);
-        }
-        // If last statement wasn't an expression, push Nil
-        self.code.emit(OpCode::LoadNil);
-    }
-}
